requirements.txt
kalshi-python>=2.1.0
cryptography>=46.0.0
python-dotenv>=1.0.0
